refactor: extract DWARF call-site expression parsing
Move caller-side target, parameter, and call-value lowering into
dwarf_expr::call_site so block indexing no longer owns DWARF op
parsing or call-site fallback behavior.
