+++ ghostscope-dwarf/src/dwarf_expr/mod.rs
@@ -1,5 +1,6 @@
 //! Shared DWARF expression parsing, scanning, and lowering utilities.
 
+pub(crate) mod call_site;
 pub(crate) mod const_eval;
 pub(crate) mod entry_value;
 pub(crate) mod errors;
